add base class for EuropeanDynamicsSpider, use it on zambia and ghana spider
