+++ kingfisher_scrapy/spiders/ghana.py
@@ -1,12 +1,7 @@
-import datetime
-
-import scrapy
-
-from kingfisher_scrapy.base_spider import CompressedFileSpider
-from kingfisher_scrapy.util import components, handle_http_error, join
+from kingfisher_scrapy.spiders.europe_dynamic_spider_base import EuropeDynamicSpiderBase
 
 
-class Ghana(CompressedFileSpider):
+class Ghana(EuropeDynamicSpiderBase):
     """
        Domain
          Ghana Electronic Procurement System (GHANEPS)
@@ -22,26 +17,7 @@ class Ghana(CompressedFileSpider):
     name = 'ghana'
 
     # BaseSpider
-    date_format = 'year-month'
     default_from_date = '2019-07'
 
-    # SimpleSpider
-    data_type = 'record_package'
+    url = 'https://www.ghaneps.gov.gh/ocds/services/recordpackage/getrecordpackagelist'
 
-    def start_requests(self):
-        yield scrapy.Request(
-            'https://www.ghaneps.gov.gh/ocds/services/recordpackage/getrecordpackagelist',
-            meta={'file_name': 'list.json'},
-            callback=self.parse_list,
-        )
-
-    @handle_http_error
-    def parse_list(self, response):
-        urls = response.json()['packagesPerMonth']
-        for url in reversed(urls):
-            if self.from_date and self.until_date:
-                year, month = map(int, url.rsplit('/', 2)[1:])
-                url_date = datetime.datetime(year, month, 1)
-                if not (self.from_date <= url_date <= self.until_date):
-                    continue
-            yield self.build_request(url, formatter=join(components(-2), extension='zip'))
